+++ src/utils/utils.cpp
@@ -36,7 +36,7 @@ Point ston(float x, float y) {
 }
 
 float precisionRound(float x, int precision) {
-	int mult = pow(10, precision);
+	int mult = static_cast<int>(pow(10, precision));
 	int num = static_cast<int>(x * mult + 0.5f);
 	return num / static_cast<float>(mult);
 }
